fix leaks in unset echo with no arg
