+++ builtin/cd.c
@@ -6,40 +6,11 @@
 /*   By: pruenrua <pruenrua@student.42bangkok.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/30 01:51:43 by pruenrua          #+#    #+#             */
-/*   Updated: 2024/01/12 12:07:52 by pruenrua         ###   ########.fr       */
+/*   Updated: 2024/01/12 20:01:01 by pruenrua         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/minishell.h"
-#define CDAGUMENTS "minishell: cd: too many arguments\n"
-
-char	*get_new_pwd(int status, char *old_pwd)
-{
-	char	*new_pwd;
-
-	new_pwd = getcwd(NULL, 0);
-	if (new_pwd == NULL && errno == ENOENT && status == 0)
-	{
-		perror("getcwd : ");
-		new_pwd = ft_strjoin(old_pwd, "/.");
-	}
-	return (new_pwd);
-}
-
-char	*get_old_pwd(t_tok *token)
-{
-	char	*old_pwd;
-
-	old_pwd = getcwd(NULL, 0);
-	if (old_pwd == NULL && errno == ENOENT)
-	{
-		perror("getcwd : ");
-		old_pwd = get_value_from_key("PWD", token->env_token);
-		if (old_pwd == NULL)
-			old_pwd = ft_strdup(token->pwd);
-	}
-	return (old_pwd);
-}
 
 int	ch_old_pwd(t_env *env_tok)
 {
